IDEMPIERE-455 Discover and fix FindBugs problems / Pattern ICAST_IDIV_CAST_TO_DOUBLE

+++ org.adempiere.base/src/org/compiere/print/layout/TableElement.java
@@ -284,7 +284,7 @@ public class TableElement extends PrintElement
 		m_printRows = new SerializableMatrixImpl<ArrayList<Serializable>>("TableElementPrintRows");	//	reset
 
 		//	Max Column Width = 50% of available width (used if maxWidth not set)
-		float dynMxColumnWidth = m_firstPage.width / 2;
+		float dynMxColumnWidth = m_firstPage.width / (float)2;
 
 		//	Width calculation
 		int rows = m_data.getRowCount();
